app/requirements.txt
# Core dependencies
streamlit==1.29.0
pandas==2.1.4
numpy==1.26.2
plotly==5.18.0
streamlit-aggrid
python-dotenv==1.0.0

# Database
pyodbc==5.0.1
sqlalchemy==2.0.23

# API and HTTP
requests==2.31.0
urllib3==2.1.0
aiohttp==3.9.1

# AI Integration
asyncio-throttle==1.0.2
python-json-logger==2.0.7

# Date and time
python-dateutil==2.8.2

# Configuration
pyyaml==6.0.1